inversion operator, refactor mutation
