+++ src/genetic/mutation/BoundaryFlipBit.ts
@@ -5,8 +5,6 @@ import BitFlipMutation from "./BitFilpMutation";
 
 export default class BoundaryFlip extends BitFlipMutation {
     
-
-
     mutate(chromosome: Chromosome) {
         if(!this.isMutate()) {
             return
